% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13_get_patho_class.R
\name{get_patho_class}
\alias{get_patho_class}
\title{Classify Pathological Subtypes Using a Pretrained CTransPath Model}
\usage{
get_patho_class(
  seurat,
  pretrain_path,
  save_dir,
  n_components = 50L,
  n_clusters = 6L,
  plot_classes = TRUE,
  image_save_path = NULL,
  ...
)
}
\arguments{
\item{seurat}{A Seurat object containing spatial transcriptomics data.}

\item{pretrain_path}{Path to the pretrained CTransPath model weights file (.pth).}

\item{save_dir}{Path to save results.}

\item{n_components}{Integer. Number of principal components for dimensionality reduction. Default is 50.}

\item{n_clusters}{Integer. Number of pathological clusters to identify. Default is 6.}

\item{plot_classes}{Logical. Whether to generate a plot of the classified subtypes. Default is TRUE.}

\item{image_save_path}{Path to save the classification plot. If NULL, the plot is not saved to disk. Default is NULL.}

\item{...}{Additional arguments:
\describe{
\item{seed}{Integer. Random seed for reproducibility. Default is 123.}
\item{assay}{Character. Name of the assay in the Seurat object to use. Default is "RNA".}
}}
}
\value{
Invisible TRUE upon successful completion.
}
\description{
Crops tiles, generates embeddings using CTransPath, performs PCA and K-means,
and adds the results to the AnnData object.
}
