% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_is_tirank_installed.R
\name{is_tirank_installed}
\alias{is_tirank_installed}
\title{Check if tirank Python module is available}
\usage{
is_tirank_installed(abort = FALSE)
}
\arguments{
\item{abort}{Logical. If TRUE, throw an error when tirank is not available.
If FALSE (default), return FALSE without error.}
}
\value{
Logical. TRUE if tirank module is available, FALSE otherwise.
}
\description{
Check if tirank Python module is available
}
