% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_normalize_data.R
\name{normalize_data}
\alias{normalize_data}
\title{Normalize gene expression data using z-score normalization (row-wise).}
\usage{
normalize_data(exp, ...)
}
\arguments{
\item{exp}{A matrix containing bulk expression data (counts) to be normalized}

\item{...}{Additional arguments (currently unused)}
}
\value{
Normalized expression data
}
\description{
Normalize gene expression data using z-score normalization (row-wise).
}
\details{
Input should be counts.

We don't use \code{base::scale} here because the difference
of ddof between R and Python.
}
\examples{
\dontrun{
normalized <- normalize_data(exp_matrix)
}
}
