% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_transfer_exp_profile.R
\name{transfer_exp_profile}
\alias{transfer_exp_profile}
\title{Get expression profile from Seurat object}
\usage{
transfer_exp_profile(sc_seurat, assay = "RNA", ...)
}
\arguments{
\item{sc_seurat}{A Seurat object containing single-cell expression data}

\item{assay}{Name of the assay to extract expression data from, defaults to "RNA"}

\item{...}{Additional arguments (currently unused, reserved for future extension)}
}
\value{
A data frame containing the expression matrix with genes as rows and cells as columns
}
\description{
Get expression profile from Seurat object
}
