% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_zzz.R
\docType{data}
\name{ts_cli}
\alias{ts_cli}
\title{Add timestamp to cli functions}
\format{
An object of class \code{environment} of length 4.
}
\usage{
ts_cli
}
\description{
Add timestamp to cli functions
}
\keyword{internal}
