% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_run_sc_tirank_model.R
\name{get_default_gpextractor_params}
\alias{get_default_gpextractor_params}
\title{Get default GPextractor parameters}
\usage{
get_default_gpextractor_params(user_list = NULL)
}
\arguments{
\item{user_list}{A named list of user-provided parameters to override defaults}
}
\value{
A list of parameters for GPextractor with user overrides applied
}
\description{
Get default GPextractor parameters
}
\seealso{
Other TiRank_params: 
\code{\link{get_default_model_params}()}
}
\concept{TiRank_params}
\keyword{internal}
