% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9_pickle_dump.R
\name{pickle_dump}
\alias{pickle_dump}
\title{Save R object to a Python pickle file}
\usage{
pickle_dump(x, save_path = NULL)
}
\arguments{
\item{x}{An R object to be serialized and saved. It must be transferrable to Python.}

\item{save_path}{Character string specifying the file path where the pickle file will be saved. Defaults to "./"}
}
\value{
Invisible \code{TRUE} The function is called for its side effect of saving the file
}
\description{
Save R object to a Python pickle file
}
